requirements.txt
discord.py==2.6.3
python-dotenv==1.1.1
aiosqlite==0.19.0
